refactor(api): change user admin route from /v1/auth to /v1/admin
