+++ src/qfa/api/app.py
@@ -27,7 +27,7 @@ from qfa.adapters.llm_client import LiteLLMClient
 from qfa.adapters.presidio_anonymizer import PresidioAnonymizer
 from qfa.adapters.tracking_llm import TrackingLLMAdapter
 from qfa.api.routes import router
-from qfa.api.routes_auth import router as auth_router
+from qfa.api.routes_admin import router as auth_router
 from qfa.api.routes_usage import router as usage_router
 from qfa.api.schemas import (
     ApiErrorDetail,
